Do not show pagination when only one page
If there is only one page of items, there is no reason to show anything related
to pagination.
This commit brings this feature to all DataTables in the Space Explorer. It
could be extended to other DataTables with a little copy-paste to modify each
tables config, but as long as so much grepping is involved I would rather just
refactor the way we configure DataTables to bring all common configuration
options to a central location. Each individual DataTable could then inherit from
our global DataTable config.
That, however, is best saved for another commit ;)

No files were added, nor global variables created in this commit.

http://stardev.cs.uiowa.edu:8078/browse/STAREXEC-435


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@24779 bd178f5f-7bea-4014-844f-495ca19b8805