Do not show trash icon if user is not logged in
This adds a new tag `<star:userLoggedIn>` that only outputs its body if the user
is currently logged in.
I could have instead created a tool with more fine grained control, which would
allow selecting users that had certain permissions, but that would also
introduce some potential bugs that would take work to avoid. This seemed like
the most maintainable solution going forward: it is semantically clear, and it
should be clear that something should either be avaliable to anon viewers or
only authenticated users.

http://stardev.cs.uiowa.edu:8078/browse/STAREXEC-432



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@24780 bd178f5f-7bea-4014-844f-495ca19b8805