Refactor DataTable config
This commit does two things that lay significant groundwork:

 1. Creates a `star` namespace for global JS functions and variables
 2. Creates a constructor for DataTable configurations

From a performance perspective, the `star` namespace does not actually reduce
the number of global variables that we have floating around (and global
variables in JS are BAD), but it will help to make explicit when we are using
global variables and when we are using function scoped vars.
(Remember that JavaScript does not have block scoping or even file scoping, so
anything not local to a function is global)

The constructor will take a dictionary as a parameter and merge it with a
default, site-wide DataTable configuration. Anything in the new dictionary will
override the defaults. Our overrides should be as limited as possible, so that
DataTables will behave consistantly throughout the site (which is not currently
the case)


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@24847 bd178f5f-7bea-4014-844f-495ca19b8805