Group actions with Primitives
Instead of stuffing all of the buttons in a "junk drawer" at the bottom of the
page, group them with the primatives they relate to. This should make it easy
enough to find a desired action that there is no longer any need for an
"advanced actions" drawer in the Space Explorer. Simplicity begets simplicity.

The order of buttons has not changed, and could possibly use some tweaking.

http://stardev.cs.uiowa.edu:8078/browse/STAREXEC-437


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@24853 bd178f5f-7bea-4014-844f-495ca19b8805