Pagination buttons should match theme
Avaliable buttons now have more contrast to be more obviously clickable.

When hovering hover a button, it will now have the same orange background as
when hovering over an element in the tree view.

Current page is still black number in a white box, but the subtle gradient of
the background has been disgarded to match the "flat" look of the rest of the
site.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@24889 bd178f5f-7bea-4014-844f-495ca19b8805