Group actions with Primitives
This extends the work done in r24853 to the recycle bin. Long term, we should
probably implement "select all/none" on this page, instead of having buttons for
the special case of deleting or restoring all.
The pagination on this page needs to be updated anyway, so both of those can be
done together in the future.

http://stardev.cs.uiowa.edu:8078/browse/STAREXEC-437



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@24893 bd178f5f-7bea-4014-844f-495ca19b8805