Upgrade to Gson 2.8.0
The feature that I'm most excited about (introduced in 2.4) is the ability to
call JsonArray.add() with a String and have Gson take care of converting the
String to a JsonElement. This will allow our code to be a lot more readable.

In fact, there are places in our code where we first create a large array of
String, and then Create a new array of JsonElement and loop through the arrays
converting items one by one. We can now eleminate this unnecessary extra loop.

All of our current integration tests pass, so it seems there are no breaking
changes here.

https://github.com/google/gson


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@24979 bd178f5f-7bea-4014-844f-495ca19b8805