Improve performance of Job Details page
The Job Details page should now load substantially faster for jobs that include
tens of thousands of job pairs.

The aim here was to try to minimize the amount of memory it took to render the
page. I tried to do this by creating static templates for strings and DOM nodes
that we were using thousands of times.

The page can still cause Chrome to hang for a few seconds, but now this happens
when the page is first loaded, and everything is like butter after just a few
seconds. Previously, it would take nearly a minute (!) before the page was
responsive.

https://stardev.cs.uiowa.edu:8078/browse/STAREXEC-447



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25111 bd178f5f-7bea-4014-844f-495ca19b8805