Display external link icon next to links in Solver Summary table
I think this is much better done in pure CSS than by adding an image tag to each
and every link in the tabe. This preserves seperation of content and
presentation, and is faster for the browser to render.

I would to start using this style everywhere we include this icon, especially if
it means we can stop including it in JSON responses, where it _really_
contributes to huge download sizes.

### On a personal note

I'm really not much a fan of `target="_blank"`

It is not difficult for a user to open a link in a new window if so desiered.
One can right click a link and select "Open in new tab" or can Ctrl-Click a link
and it will open in a new tab. The opposite is far more difficult: if a link is
hard-coded to open in a new tab, a user cannot easily decide to open that link
in the current tab.

This can lead to confusing History stacks so that when a user clicks the Back
button, they might not get back to where they expected.

All that said, this has been the behavior for quite some time, and I have no
intention of breaking that simply for my own preferences.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25116 bd178f5f-7bea-4014-844f-495ca19b8805