jobAttributes page broken?
Things were not loading properly due to the order in which the scripts were
added to the page; scripts are exectued in the order they appear in the HTML.
Anything we require should load BEFORE the file that requires it.

I have fixed the error messages in the console, but the page still isn't
displaying any real data. I'm not sure what it is *supposed* to be displaying so
this is a difficult problem for me to debug.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25119 bd178f5f-7bea-4014-844f-495ca19b8805