Disable hiding Select All buttons on Create Job page
The selection tables on the Create Job page do not behave in the same way as
selectable tables in the Space Explorer. The Select All button was being hidden
from Solver Selection if there was only one Benchmark available to select.

While it would be nice to hide the Select All button from the Benchmark
Selection screen if there is only one Benchmark, that should not cause all
Select All buttons to be hidden.

For now, I have restricted the Select All hiding to just instances like the
Space Explorer where the selection table is in an expandable fieldset.

There are plans to completely change the selection process in the future, and I
would very much like to rework the Create Job workflow as well.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25131 bd178f5f-7bea-4014-844f-495ca19b8805