Add five-thousand-configs solver
This does not work!
http://stardev.cs.iuowa.edu:8078/browse/STAREXEC-453

This is the solver that I used to stress test the Job Details page.
It contains 5,000 configs. I ran all configs against 10 benchmarks to create a
job with 50,000 job pairs.

I created a `.tar.gz` archive because it was the most efficient way to compress
5,000 identical files. (Zip file was 20x the size.) Problem is, StarexecCommand
does not currently supporting anything other than a Zip archive.

Until we can fix that, you can still use this solver for testing, you just need
to upload it through the web interface.



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25139 bd178f5f-7bea-4014-844f-495ca19b8805