killAjaxRequests before requests that will load the page
If the user is triggering a refresh of the page anyway, there is no need to wait
for anything on the current page to be updated.

This commit does not go far enough.
One of the biggest sources of AJAX requests on the Job Details page is from
DataTables that are set to continuously refresh. The AJAX requests generated by
DataTables are not added to our `openAjaxRequests` queue, so we have no handle
with which to cancel them.

In some extreme instances, the page can create DataTables refresh requests
faster than they can be completed. When this happens the page will seem to stop
responding, the browser may refuse to open any new connections to StarExec, and
it may even become impossible to open a StarExce page in a different browser
tab.

These DataTables refreshes are now the most significant performance bottleneck
on the Job Details page. We need to find some way to minimize them and/or keep
track of them in `openAjaxRequests` so we can cancel them if necessary.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25266 bd178f5f-7bea-4014-844f-495ca19b8805