isValidUploadRequest should allow any allowed archive type to be uploaded
Why VALID_ARCHIVETYPES does contain all allowed archive types I cannot say. I
also do not feel comfortable removing VALID_ARCHIVETYPES because it is used in
CommandParser.pollJob and I do not know if it is intended for only ZIP files to
be used in that context or any allowed archive.

Matters are further complicated by duplicated code. Identical declerations of
`extensions` can be found in:
  * servlets/UploadSolver.java
  * servlets/BenchmarkUploader.java
  * servlets/ProcessorManager.java
  * servlets/UploadSpaceXML.java
  * servlets/UploadJobXML.java



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25268 bd178f5f-7bea-4014-844f-495ca19b8805