Refactor `expandable`
Previously, we allowed a callback function to be passed to `expandable` which
would be called when a fieldset was opened (but not when it was closed). This
was a fairly limiting design, and is currently not used _anywhere_ in our
codebase.

This adds Events which will be triggered when a fieldset is opened or closed.

I hope to use this to defer loading of tables in a collapsed fieldset.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25297 bd178f5f-7bea-4014-844f-495ca19b8805