Fix for older browsers
In some older browsers (including Firefox 45.2.0) assigning a native function to
a `var` throws `TypeError`. To work around this, this commit creates a new
instance that is still bound to the `console` object instead of the global
object. There is no downside to doing this, the browser console still gives us
the correct line numbers when messages are logged (unlike before, when we had
created our own logging function).

I would just like to note that Firefox 52 never had any problem with the old
code.

Simple fix, once we figured out exactly what the problem was :+1:


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25446 bd178f5f-7bea-4014-844f-495ca19b8805