`processPairsToSolverStats` can return Collection instead of List
All we ever do with the return from `processPairsToSolverStats` is iterate
through it and get its size. There is no reason this needs to be a List, it can
be a Collection instead.

If we allow this to be a Collection, than we can return `Hashtable.values()`
directly instead of building a List first, and nothing else needs to change.



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25455 bd178f5f-7bea-4014-844f-495ca19b8805