Do not create Subspace Summaries until requested
Subspace Summaries can be quite large, and loading them can cause other parts of
the page to become unresponsive. This commit defers creating these tables until
the container they are inside is expanded.

Because JavaScript is a Function Scoped language, I had to do a little dancing
around to get the lambda function to be created with the appropriate scope.
Without wrapping the body of the `for` loop inside a function, `$panel` and
`panelTableInitializer` would always refer to the values they contained during
the *last* iteration of the loop, rather than during the iteration of the loop
that created the lambda function as necessary.

There are certainly other ways of doing this, and likely a better way of doing
this, but I think simply wrapping the body of the loop in a function results in
the most readable code, and has the benefit of behaving how a Java developer
would expect it to behave. This seems like a better compromise than throwing
function parameters back and forth with `bind` magic.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25463 bd178f5f-7bea-4014-844f-495ca19b8805