Introduce scaffolding for Analytics feature
This provides the database backend for the Analytics feature.

Still needed is the Java class to call the procedure, and we will probably want
to add more than one action before we deploy this feature, though I think
starting small and adding more actions later would be the best idea.

The concept here is nearly exactly what Albert proposed in
http://stardev.cs.uiowa.edu:8078/browse/STAREXEC-459


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25599 bd178f5f-7bea-4014-844f-495ca19b8805