I don't know why I was using that underscore thing.
The deal is, since `this.id` is final, Java needs to be sure that we are setting
it once and only once. For this reason, we need to jump through a bunch of hoops
to set this temporary `id` variable inside the `try` block, and only set
`this.id` at the very end.

At any rate, shadowing `id` locally is just fine; there's no reason to call it
`_id`.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25619 bd178f5f-7bea-4014-844f-495ca19b8805