Record QuickJob use via Analytics
Is this a popular feature?

This will only record when a Quick Job is actually created; it will not track
pageviews of the Quick Job page. It might be useful to track that additionally
to see if users are finding that page but not going through with using the
feature.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25625 bd178f5f-7bea-4014-844f-495ca19b8805