Get Analytics data for specified range
This procedure will return all Analytics data for any events recorded during the
(inclusive) date range. I hope this flexability allows for getting records for a
given week or month.

The date in the results corresponds to the start date. If displaying results for
a given week, I hope this can be used to create a label like "For the week of
Apr 16 2017" or something similar.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25690 bd178f5f-7bea-4014-844f-495ca19b8805