Begin frontend for Analytics system
This is a very basic start for a frontend to the Analytics system.

Right now, it simply displays a table with count totals for all events between
two dates. On any browser aside from Chrome and Edge, the date needs to be input
manually using the format yyyy-MM-dd. There is a jQuery UI widget that will
allow this to work in those browsers; I will implement that soon.

Right now there is no filtering for specific events, or any way to see how the
count changes over time. I would like to implement that soon.

Keeping in mind that this system is only ever going to be accessed by developers
and admins, I don't think these compromises are terrible for a first start.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25783 bd178f5f-7bea-4014-844f-495ca19b8805