Use RegEx to extractIDFromJson
Parsing HTML with regular expressions is a bad idea, but trying to
reinvent regular expressions is a worse idea. Eventually, it would be
nice to include the information we need directly in the JSON instead of
trying to extract it from strings of HTML in the JSON. It would be
trivial to include both, and we could make our JSON blobs smaller and
faster if we stopped transmitting strings of HTML completely and just
let the browser build what it needs dynamically.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25892 bd178f5f-7bea-4014-844f-495ca19b8805