Add variables to StarExecCommand
TODO Before Deploy:
 * Add documentation
 * Change StarExecCommand version ID


This adds basic support for variables to StarExecCommand.
This will aid in writing scripts for StarExecCommand that do things like
create a new subspace and then upload primitives to that subspace.
Previously, the ID of the new subspace had to be hardcoded in the
script, which made this kind of scripting effectively impossible for
anything other than a fresh install of StarExec. Now, the ID of the new
subspace can be assigned to a variable, and that variable can be used as
the ID of the space for the uploaded primitives.

Because I wanted to do this in the simplest way possible without
breaking things, there are some weird edge cases. If a variable is
uninitialized, it will act as a string constant instead of a variable.
That is to say
 `$uninitialized`
will evaluate to
 `$uninitialized`
This is because there is no rule against strings starting with a '$', so
we cannot assume that anything starting with a '$' is in fact a
variable.

A possible workaround would be to only allow variables for parameters
that expect a number, but that would require a good deal of refactoring
to how CommandParser works. (CommandParser could probably use some
refactoring.)



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25899 bd178f5f-7bea-4014-844f-495ca19b8805