Changes SUCCESS codes
We want to be able to return primitive IDs without collisions.

Previously, when doing a new install, a space would be created and
return an ID of `1`, which the Shell would interpret as the exit
sentinel and quit. Now, so long as a StarExec instance never has more
than 2^31-5 primitives we should be safe.

It would be nice if we could use enums or algebraic data types to pass
status back and forth, but this fixes the problem for now.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25915 bd178f5f-7bea-4014-844f-495ca19b8805