Debug messages should throw exceptions
Prior to r24491 this would just silently log null and move on with life.
I don't have any idea what this debug message was trying to debug. It
was added back in r18275. Since we are no longer using `debugSeries` or
`debugItem` anywhere, we could probably just get rid of this debug
statement..?


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25928 bd178f5f-7bea-4014-844f-495ca19b8805