Force to Dates from same day to be equal
According to the Java documentation, the constructor for java.sql.Date
should remove any time information and change the time to midnight GMT.
This does not seem to be the case however. This means that two Dates
from the same day were not equal, which totally defeated the purpose.

This change manually zeros out the time. There *must* be a better way to
do this, but this is the only method I have found that also zeros out
the milliseconds.

This took me too long to debug. 😓

https://docs.oracle.com/javase/8/docs/api/java/sql/Date.html#Date-long-




git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@26124 bd178f5f-7bea-4014-844f-495ca19b8805