Don't package everything before unit tests
If we only want to run tests, there is no need to package everything for
deployment. This allows `ant test` to run in about half the time of what
`ant epilog-test` takes.

The target "epilog-test" behavior is unchanged, though "package" or
something similar might be a better name at this point.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@26291 bd178f5f-7bea-4014-844f-495ca19b8805