Enable client-side debugMode for developers
This is better than turning it on or off globally for all users at build
time. Better still would be to hide an option somewhere that can turn it
on or off via a cookie on a users machine. This way, if we needed a user
to debug something for us (unlikely) we could guide them through turning
debugMode on and sending us the browser's JS log.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@26294 bd178f5f-7bea-4014-844f-495ca19b8805