Properly log StarExecCommand logins
This property has existed in Analytics for a while, but never worked as
expected. Instead of relying on an empty header property which is never
detected, this sets the user-agent of StarExecCommand, and matches that
UA string.

I would really like to include the build number in that UA string for
forward-compatibility, but it is not available anywhere when building
StarExecCommand. It will require changes to the build system, and is not
the highest priority at the moment.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@26351 bd178f5f-7bea-4014-844f-495ca19b8805