Move JobStatus tests from JUnit to Integration
This information can and should be computed by the database rather than
in Java. As a step toward that, I am rewrote the tests as integration
tests so that they will have access to stored procedures and the like.

I am doing this step FIRST so as to not make dramatic changes to logic
and tests in the same commit.

The new tests are passing as of this commit.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@26435 bd178f5f-7bea-4014-844f-495ca19b8805