Quiet logging of bad requests.
If a user requests the QuickJob page without providing a Space ID, they
are given an error message. This error message is relevant to *them* but
not so much to *us*. This is not an error that we need to fix, it is
just an ill-formed request from the user.

This information might (possibly, at some point) be informative while
debugging, but it does not need to be included in the regular emailed
error logs.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@26445 bd178f5f-7bea-4014-844f-495ca19b8805