Add a solver that has a binary config instead of shell script
This was not easy to create, but I swear I've seen somebody on StarExec
doing something similar.

I cannot figure out how to upload the binary executable directly in the
zip file; the program needs to be built on StarExec or it will not run:

> cannot execute binary file

With this solver, I will work on disallowing view-source for binary
configs.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@26544 bd178f5f-7bea-4014-844f-495ca19b8805