If no sandbox can be found, immediately terminate
This patch does not actually change any behavior. It just makes explicit
that the jobpair will terminate immediately if no sandbox can be
secured. There is thereafter never and need to make sure that there is a
sandbox; there will never be a state where the script is running without
a sandbox.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@26826 bd178f5f-7bea-4014-844f-495ca19b8805