Don't bother querying for null processor
It is kind of unavoidable that we will sometimes call `get` when there
is no processor we are actually looking for. In these cases, just
immediately return null. Otherwise, we will add many lines of SQL
exception noise to the log file for no good reason.

In a worst case scenario, this could cause a stressed out developer to
go looking in all the wrong places when trying to hunt down a loosely
related bug ☠︎




git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@26977 bd178f5f-7bea-4014-844f-495ca19b8805