Client closing connection is not a bug
This method has been producing huge stack traces in weekly error logs.
The client closing the connection is (probably) not a bug in our code.
Rather than the stack trace, it would be more helpful to simply have the
URL the user was trying to access. This will make it easier to figure
out what the user was trying to do before the connection was closed. If
this turns out to be benign, it should be downgraded from `warn` to
`debug`.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@27634 bd178f5f-7bea-4014-844f-495ca19b8805