Split SQL into own build file
I would like to move to deploying SQL updates via Ant rather than via
shell script. Ant can take care of a lot of the messiness for us. This
change lays the groundwork for that, but it doesn't actually work yet
because this requires some changes to our SQL syntax that are
incompatible with the current deploy process.

When we change our deploy process to use Git instead of SVN, I think we
could also change to update SQL via Ant instead of shell script.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@28801 bd178f5f-7bea-4014-844f-495ca19b8805