Move default properties out of Ant XML into properties file
It should be easy to look up what the defaults are, and these should be
all in one place. Where it is avoidable, we should avoid mixing these
properties with the Ant XML build files.

The only properties that should be set via Ant are build-time dynamic
properties, like the current time or the revision number.


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@28929 bd178f5f-7bea-4014-844f-495ca19b8805