Avoid "use starexec"
We should not hardcode the database name in our SQL files, even if we do
use a regex to update it at build time. The database name can be
provided as an argument when using the `mysql` command. Furthermore, by
not including "use starexec" we can use Ant to execute SQL scripts for
us. This will greatly aid updating SQL procedures during deploy.

Do we need to keep historic SQL changes around?


git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@28952 bd178f5f-7bea-4014-844f-495ca19b8805