Log Python API logins
If we ever deployed a theoretical Python API, it would be nice to have
some way to measure its usage (as we already do for StarExecCommand).

This change assumes that any request from this theoretical Python API
will contain "starexec.py" in the `User-Agent`. The `User-Agent` would
likely contain other information as well, such as the version of Python
or the name of the program accessing the Python API bindings, but for
now all we care about is this simple check.