Use `umask` to set file permissions (#201)
We must ensure that any directories created with `mkdir -p` are readable
by the group `star-web`. By setting `umask` at the beginning of the
script, we ensure that any directories created in this context will be
group read+write+execute. The `umask` is limited to the scope of the
script, and will revert to what it was before after the script
terminates.

Unix is fun.