small tweak to overcome issue copying files to sandbox directory
New temporary directories seem to get created without group write
permission as of the tc9 upgrade (based on comments from Alex in the
code), and the code is counting on having that permission.  Simple
solution is to chmod to add group-write permission.  Alex already had
a function in there to do that, so we just call it.